[project @ 2002-01-29 14:41:52 by simonmar] Fix highly obscure bug.
The heap check in cmpIntegerzh_fast was erroneously passing
cmpIntegerIntzh_fast as the resumption point instead of
cmpIntegerzh_fast, with the result that if a heap-check happened to
strike in cmpIntegerzh_fast then we would end up resuming in
cmpIntegerIntzh and returning an incorrect comparison result.  One
symptom is that very occasionally floating point numbers would print
incorrectly (ending in an 'a').