[project @ 2002-01-30 17:16:36 by simonpj] ----------------------------- 	Tidy up the top level of TcModule 	-----------------------------
This commit started life as sorting out the TcInstDcls thing that
we got wrong a few weeks back, but it spiraled out of control.

However, the result is a nice tidy up of TcModule.

typecheckModule/tcModule compiles a module from source code
typecheckIface/tcIface   compiles a module from its interface file
typecheckStmt		 compiles a Stmt
typecheckExpr		 compiles a Expr

tcExtraDecls is used by typecheckStmt/typecheckExpr
	to compile interface-file decls.
	It is just a wrapper for:

tcIfaceImports, which is used by tcExtraDecls and tcIface
	to compile interface file-file decls.

tcImports, is similar to tcIfaceImports, but is used only by tcModule

tcIfaceImports is used when compiling an interface, and can
	therefore be quite a bit simpler