[project @ 2002-03-01 18:28:15 by keithw] Addendum to previous commit message:
We can't tell whether the stg_sel and stg_ap thunks correspond to
static or dynamic thunks, and it's not worth duplicating them all just
so the ticky-ticky stats can be correct.  So I have played it safe
(for my purposes at least!) and counted them all as dynamic.  Yell if
this is a problem (njn?)