[project @ 2002-03-02 18:02:30 by sof] Urk, code generator assumed reversed info tables in its implementation of dataToTag# - broke a compiler built unregisterised rather weirdly (and only, none of the code in testsuite/ nor nofib/ showed up this bug.)
This commit assumes that unregisterised==
info_tables_not_next_to_code & calls upon
a trusty old PrimOps.h macro to locate the tag
in the info table (the reasons for doing it
this way is explained in AbsCUtils.lhs
comments).

To help the poor sod who has to debug a break
like this sometime in the future, I'm trying to
come up with a repro case smaller than
ghc/compiler/ -- not yet successful.

Anyway, this concludes the fixes to the
unregisterised bits; I hereby claim that it
is now working again.