[project @ 2002-03-08 15:47:18 by simonpj] ------------------------ 	Kill Type.splitRepFunTys 	------------------------
splitRepFunTys was a Bad Function that split up a function type
looking through even recursive newtypes.  Alas, it diverged if
when we had a recursive newtype with a function whose result was
the newtype itself.

I've replaced it with ordinary splitFunTys, plus a new function
Type.dropForAlls, which does what you would expect.