[project @ 2002-03-27 16:48:47 by simonmar] Sigh, recompilation checking in --make mdoe has been badly broken for well, probably ever.
The problem is that in GHCi mode there's a restriction on the use of
object files, namely that an object file may only depend on other
up-to-date object files in order to be up-to-date itself, because we
don't have support for linking object code to interpreted code.
Unfortunately we were applying this restriction to --make compilations
too, with the result that far too much recompilation was happening.

MERGE TO STABLE