[project @ 2002-04-01 15:32:46 by panne] Sigbjorn's last optimization (checking for -mno-cygwin only for mingw32 targets) kicked out -O from the default SRC_CC_OPTS. Apart from a minor performance hit for some parts of GHC, it yields a GHCi which can't load HSbase_cbits.o because `lstat' is unknown, at least on SuSE 7.3.
A little investigation showed the rather arcane reason: lstat and
friends are inline functions and therefore not in libc.so, only in its
static counterpart. Normally this is not a problem at all, but the CPP
INLINE trickery in fptools/libraries/base/cbits/PrelIOUtils.c manages
to get a reference to lstat into PrelIOUtils.o if -O is not given. %-}
A similar problem exists for fstat, too.

Simple solution: Re-add -O to SRC_CC_OPTS, simplifying configure.in a
bit on the way.