[haddock @ 2002-05-08 11:21:56 by simonmar] Add a facility for specifying options that affect Haddock's treatment of the module.  Options are given at the top of the module in a comma-separated list, beginning with '-- #'.  eg.
	-- # prune, hide, ignore-exports

Options currently available, with their meanings:

  prune:
	ignore declarations which have no documentation annotations

  ignore-exports:
	act as if the export list were not specified (i.e. export
   	everything local to the module).

  hide:
	do not include this module in the generated documentation, but
 	propagate any exported definitions to modules which re-export
	them.

There's a slight change in the semantics for re-exporting a full
module by giving 'module M' in the export list: if module M does not
have the 'hide' option, then the documentation will now just contain a
reference to module M rather than the full inlined contents of that
module.

These features, and some other changes in the pipeline, are the result
of discussions between myself and Manuel Chakravarty
<chak@cse.unsw.edu.au> (author of IDoc) yesterday.

Also: some cleanups, use a Writer monad to collect error messages in
some places instead of just printing them with trace.