[haddock @ 2002-06-21 15:50:42 by simonmar] Add support for reading and writing interface files(!)
This turned out to be quite easy, and necessary to get decent
hyperlinks between the documentation for separate packages in the
libraries.

The functionality isn't quite complete yet: for a given package of
modules, you'd like to say "the HTML for these modules lives in
directory <dir>" (currently they are assumed to be all in the same
place).

Two new flags:

	--dump-interface=FILE   dump an interface file in FILE
	--read-interface=FILE	read interface from FILE

an interface file describes *all* the modules being processed.  Only
the exported names are kept in the interface: if you re-export a name
from a module in another interface the signature won't be copied.
This is a compromise to keep the size of the interfaces sensible.

Also, I added another useful option:

	--no-implicit-prelude

avoids trying to import the Prelude.  Previously this was the default,
but now importing the Prelude from elsewhere makes sense if you also
read in an interface containing the Prelude module, so Haddock imports
the Prelude implicitly according to the Haskell spec.