[project @ 2002-09-27 08:16:24 by simonpj] --------------------------------      Do type-checking of external-core input 	--------------------------------
When we read in an External Core file, we should really type-check it.
We weren't, because we treated it as if it were trusted, interface-file
material.

This commit fixes the problem, albeit in a bit of a hacky way.  The
typechecking is done by Lint, which does not give as friendly error
messages as does the normal typechecker.  But it's much better than nothing.

I also removed the entirely-unused 'warnings' from the Lint monad.